+++ team_token_integration_test.go
@@ -6,6 +6,7 @@ package tfe
 import (
 	"context"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -39,6 +40,55 @@ func TestTeamTokensCreate(t *testing.T) {
 		assert.Equal(t, err, ErrInvalidTeamID)
 	})
 }
+
+func TestTeamTokens_CreateWithOptions(t *testing.T) {
+	client := testClient(t)
+	ctx := context.Background()
+
+	tmTest, tmTestCleanup := createTeam(t, client, nil)
+	defer tmTestCleanup()
+
+	var tmToken string
+	t.Run("with valid options", func(t *testing.T) {
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		tmToken = tt.Token
+	})
+
+	t.Run("when a token already exists", func(t *testing.T) {
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		assert.NotEqual(t, tmToken, tt.Token)
+	})
+
+	t.Run("without valid team ID", func(t *testing.T) {
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, badIdentifier, TeamTokenCreateOptions{})
+		assert.Nil(t, tt)
+		assert.Equal(t, err, ErrInvalidTeamID)
+	})
+
+	t.Run("without an expiration date", func(t *testing.T) {
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		assert.Empty(t, tt.ExpiredAt)
+		tmToken = tt.Token
+	})
+
+	t.Run("with an expiration date", func(t *testing.T) {
+		start := time.Date(2024, 1, 15, 22, 3, 4, 0, time.UTC)
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{
+			ExpiredAt: &start,
+		})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		assert.Equal(t, tt.ExpiredAt, start)
+		tmToken = tt.Token
+	})
+}
+
 func TestTeamTokensRead(t *testing.T) {
 	client := testClient(t)
 	ctx := context.Background()
@@ -52,6 +102,19 @@ func TestTeamTokensRead(t *testing.T) {
 		tt, err := client.TeamTokens.Read(ctx, tmTest.ID)
 		require.NoError(t, err)
 		assert.NotEmpty(t, tt)
+
+		ttTestCleanup()
+	})
+
+	t.Run("with an expiration date passed as a valid option", func(t *testing.T) {
+		start := time.Date(2024, 1, 15, 22, 3, 4, 0, time.UTC)
+
+		_, ttTestCleanup := createTeamTokenWithOptions(t, client, tmTest, TeamTokenCreateOptions{ExpiredAt: &start})
+
+		tt, err := client.TeamTokens.Read(ctx, tmTest.ID)
+		require.NoError(t, err)
+		assert.NotEmpty(t, tt)
+		assert.Equal(t, tt.ExpiredAt, start)
 
 		ttTestCleanup()
 	})